core: fix dlopen interception of libusb-1.0 functions
This commit fixes a bug introduced in 0.1.6rc1 that caused a linking
error on looking for libusb_control_transfer. This function is called
in two inline functions from libusb.h: libusb_get_descriptor, and
libusb_get_string_descriptor. Because the call is in an inline
function the #define's that overwrite the libusb_* calls do not
apply. I copied these functions into libusb-dload.h to fix the
bug. This is not a clean solution but will suffice for 0.1.6.

Signed-off-by: Nathan Hjelm <hjelmn@mac.com>