autoconf: change SONAME detection source from ldd to objdump
ldd detects used libraries by partially running the target binary. This
is a problem for when the target binary is not runnable
(e.g. cross-compilation).

Switch to objdump which only reads info from the ELF file without
running anything from it.

Note: the output of objdump -p looks like this:
  [...]
  Dynamic Section:
    NEEDED               libusb-1.0.so.0
    NEEDED               libc.so.6
  [...]